server/requirements.txt
asgiref==3.8.1
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
Django==3.2.25
django-cors-headers==4.0.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.7.0
firebase-admin==6.1.0
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
msgpack==1.1.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pymongo==3.12.3
pyparsing==3.2.3
python-dotenv==1.0.0
pytz==2025.2
requests==2.32.3
rsa==4.9.1
setuptools==80.7.1
sqlparse==0.2.4
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
